refactor(header-skin): 皮肤代码拆分为一个皮肤一个文件(part/part of),主文件只留模型+HSL工具+注册表
