+++ lib/styles/header_skins.dart
@@ -4,16 +4,25 @@ import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import '../l10n/app_localizations.dart';
 
-/// 头部皮肤系统(PoC):皮肤 = 叠在「主题色底」之上的装饰层。
+part 'header_skins/aurora_skin.dart';
+part 'header_skins/bokeh_skin.dart';
+part 'header_skins/clouds_skin.dart';
+part 'header_skins/image_skin.dart';
+part 'header_skins/mountains_skin.dart';
+part 'header_skins/pattern_skins.dart';
+part 'header_skins/sunset_skin.dart';
+part 'header_skins/waves_skin.dart';
+
+/// 皮肤系统:皮肤 = 叠在「主题色底」之上的装饰层(渲染在 PrimaryHeader)。
 ///
 /// 设计原则:**皮肤跟随用户主题色(theme-tinted)** —— 用 HSL 从 primary 派生出
 /// 渐变 / 图形,所以任何主题色都成立(「主题色 + 皮肤 = PrimaryHeader」)。
 /// - 亮色模式:整体保持在主题色的明度区间(偏亮),让 header 现有的深色文字仍可读。
 /// - 暗色模式:纯黑底(不叠主题色底),图形仍用主题色但更淡,保持白色文字可读。
 ///
-/// 两类皮肤:**代码皮肤**(渐变/几何/光斑,跟随主题色)与 **图片皮肤**
-/// (`_ImageSkin`,SVG 全幅铺满;themed=true 整幅染成主题色,否则用 SVG 自带配色)。
-/// 新增图片皮肤见 assets/header_skins/README.md。
+/// 两类皮肤:**代码皮肤**(渐变/几何/光斑,跟随主题色,见各 `header_skins/*_skin.dart`
+/// part)与 **图片皮肤**(`_ImageSkin`,SVG 全幅铺满;themed=true 整幅染成主题色,否则
+/// 用 SVG 自带配色)。**一个皮肤一个 part 文件**;新增见 assets/header_skins/README.md。
 
 const String kHeaderSkinNone = 'none';
 
@@ -33,7 +42,7 @@ class HeaderSkin {
   final Widget Function(Color primary, bool isDark) builder;
 }
 
-// ---- HSL 派生工具 ----
+// ---- HSL 派生工具(供各皮肤 part 共用)----
 Color _lighten(Color c, double amount) {
   final h = HSLColor.fromColor(c);
   return h.withLightness((h.lightness + amount).clamp(0.0, 1.0)).toColor();
@@ -99,587 +108,4 @@ HeaderSkin? headerSkinById(String id) {
     if (s.id == id) return s;
   }
   return null;
-}
-
-// ============================ 极光(渐变 + 柔光斑) ============================
-
-class _AuroraSkin extends StatelessWidget {
-  const _AuroraSkin(this.primary, this.isDark);
-  final Color primary;
-  final bool isDark;
-
-  @override
-  Widget build(BuildContext context) {
-    final colors = isDark
-        ? [Colors.black, Colors.black]
-        : [_lighten(primary, 0.20), primary, _lighten(_hueShift(primary, 30), 0.12)];
-    return DecoratedBox(
-      decoration: BoxDecoration(
-        gradient: LinearGradient(
-          begin: Alignment.topLeft,
-          end: Alignment.bottomRight,
-          colors: colors,
-        ),
-      ),
-      child: CustomPaint(
-        painter: _AuroraPainter(primary, isDark),
-        child: const SizedBox.expand(),
-      ),
-    );
-  }
-}
-
-class _AuroraPainter extends CustomPainter {
-  _AuroraPainter(this.primary, this.isDark);
-  final Color primary;
-  final bool isDark;
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    final blobs = <(Offset, double)>[
-      (Offset(size.width * 0.15, size.height * 0.25), size.width * 0.24),
-      (Offset(size.width * 0.88, size.height * 0.12), size.width * 0.18),
-      (Offset(size.width * 0.72, size.height * 0.85), size.width * 0.28),
-    ];
-    for (final (center, r) in blobs) {
-      final paint = Paint()
-        ..color = (isDark ? primary : Colors.white)
-            .withValues(alpha: isDark ? 0.13 : 0.22)
-        ..maskFilter = const MaskFilter.blur(BlurStyle.normal, 26);
-      canvas.drawCircle(center, r, paint);
-    }
-  }
-
-  @override
-  bool shouldRepaint(covariant _AuroraPainter old) =>
-      old.primary != primary || old.isDark != isDark;
-}
-
-// ============================ 山峦(低多边形 + 日月) ============================
-
-class _MountainsSkin extends StatelessWidget {
-  const _MountainsSkin(this.primary, this.isDark);
-  final Color primary;
-  final bool isDark;
-
-  @override
-  Widget build(BuildContext context) {
-    final colors = isDark
-        ? [Colors.black, Colors.black]
-        : [_lighten(primary, 0.24), _lighten(primary, 0.06)];
-    return DecoratedBox(
-      decoration: BoxDecoration(
-        gradient: LinearGradient(
-          begin: Alignment.topCenter,
-          end: Alignment.bottomCenter,
-          colors: colors,
-        ),
-      ),
-      child: CustomPaint(
-        painter: _MountainsPainter(primary, isDark),
-        child: const SizedBox.expand(),
-      ),
-    );
-  }
-}
-
-class _MountainsPainter extends CustomPainter {
-  _MountainsPainter(this.primary, this.isDark);
-  final Color primary;
-  final bool isDark;
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    final w = size.width, h = size.height;
-
-    // 日 / 月
-    final orb = Paint()
-      ..color = (isDark ? primary : Colors.white)
-          .withValues(alpha: isDark ? 0.22 : 0.5);
-    canvas.drawCircle(Offset(w * 0.78, h * 0.30), h * 0.16, orb);
-
-    // 远山
-    final back = Paint()
-      ..color = (isDark ? primary : _lighten(primary, -0.02))
-          .withValues(alpha: isDark ? 0.15 : 0.45);
-    final p1 = Path()
-      ..moveTo(0, h)
-      ..lineTo(0, h * 0.72)
-      ..lineTo(w * 0.26, h * 0.50)
-      ..lineTo(w * 0.52, h * 0.74)
-      ..lineTo(w * 0.78, h * 0.48)
-      ..lineTo(w, h * 0.70)
-      ..lineTo(w, h)
-      ..close();
-    canvas.drawPath(p1, back);
-
-    // 近山
-    final front = Paint()
-      ..color = (isDark ? primary : _lighten(primary, -0.10))
-          .withValues(alpha: isDark ? 0.26 : 0.62);
-    final p2 = Path()
-      ..moveTo(0, h)
-      ..lineTo(0, h * 0.88)
-      ..lineTo(w * 0.18, h * 0.68)
-      ..lineTo(w * 0.40, h * 0.90)
-      ..lineTo(w * 0.62, h * 0.66)
-      ..lineTo(w * 0.85, h * 0.88)
-      ..lineTo(w, h * 0.74)
-      ..lineTo(w, h)
-      ..close();
-    canvas.drawPath(p2, front);
-  }
-
-  @override
-  bool shouldRepaint(covariant _MountainsPainter old) =>
-      old.primary != primary || old.isDark != isDark;
-}
-
-// ============================ 光斑(渐变 + 散落气泡) ============================
-
-class _BokehSkin extends StatelessWidget {
-  const _BokehSkin(this.primary, this.isDark);
-  final Color primary;
-  final bool isDark;
-
-  @override
-  Widget build(BuildContext context) {
-    final colors = isDark
-        ? [Colors.black, Colors.black]
-        : [
-            _lighten(_hueShift(primary, -25), 0.12),
-            primary,
-            _lighten(_hueShift(primary, 35), 0.10),
-          ];
-    return DecoratedBox(
-      decoration: BoxDecoration(
-        gradient: LinearGradient(
-          begin: Alignment.topLeft,
-          end: Alignment.bottomRight,
-          colors: colors,
-        ),
-      ),
-      child: CustomPaint(
-        painter: _BokehPainter(primary, isDark),
-        child: const SizedBox.expand(),
-      ),
-    );
-  }
-}
-
-class _BokehPainter extends CustomPainter {
-  _BokehPainter(this.primary, this.isDark);
-  final Color primary;
-  final bool isDark;
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    final rnd = math.Random(7); // 固定种子,保持稳定
-    final palette = isDark
-        ? [primary, _hueShift(primary, 30), _lighten(primary, 0.2)]
-        : [Colors.white, _lighten(primary, 0.25), _lighten(_hueShift(primary, 35), 0.15)];
-    for (int i = 0; i < 9; i++) {
-      final x = rnd.nextDouble() * size.width;
-      final y = rnd.nextDouble() * size.height;
-      final r = size.width * (0.04 + rnd.nextDouble() * 0.14);
-      final color = palette[rnd.nextInt(palette.length)];
-      final paint = Paint()
-        ..color = color.withValues(
-            alpha: (isDark ? 0.06 : 0.08) + rnd.nextDouble() * 0.12);
-      if (i % 3 == 0) {
-        paint.maskFilter = const MaskFilter.blur(BlurStyle.normal, 12);
-      }
-      canvas.drawCircle(Offset(x, y), r, paint);
-    }
-  }
-
-  @override
-  bool shouldRepaint(covariant _BokehPainter old) =>
-      old.primary != primary || old.isDark != isDark;
-}
-
-// ============================ 波浪(渐变 + 底部叠浪) ============================
-
-class _WavesSkin extends StatelessWidget {
-  const _WavesSkin(this.primary, this.isDark);
-  final Color primary;
-  final bool isDark;
-
-  @override
-  Widget build(BuildContext context) {
-    final colors = isDark
-        ? [Colors.black, Colors.black]
-        : [_lighten(primary, 0.18), primary];
-    return DecoratedBox(
-      decoration: BoxDecoration(
-        gradient: LinearGradient(
-          begin: Alignment.topCenter,
-          end: Alignment.bottomCenter,
-          colors: colors,
-        ),
-      ),
-      child: CustomPaint(
-        painter: _WavesPainter(primary, isDark),
-        child: const SizedBox.expand(),
-      ),
-    );
-  }
-}
-
-class _WavesPainter extends CustomPainter {
-  _WavesPainter(this.primary, this.isDark);
-  final Color primary;
-  final bool isDark;
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    final w = size.width, h = size.height;
-    for (int i = 0; i < 3; i++) {
-      final baseY = h * (0.58 + i * 0.13);
-      final amp = h * 0.07;
-      final phase = i * 1.1;
-      final color = (isDark ? primary : Colors.white)
-          .withValues(alpha: isDark ? 0.07 + i * 0.03 : 0.10 + i * 0.05);
-      final path = Path()..moveTo(0, baseY);
-      for (double x = 0; x <= w; x += w / 48) {
-        path.lineTo(x, baseY + amp * math.sin((x / w) * 2 * math.pi * 1.4 + phase));
-      }
-      path
-        ..lineTo(w, h)
-        ..lineTo(0, h)
-        ..close();
-      canvas.drawPath(path, Paint()..color = color);
-    }
-  }
-
-  @override
-  bool shouldRepaint(covariant _WavesPainter old) =>
-      old.primary != primary || old.isDark != isDark;
-}
-
-// ====== 几何图案皮肤(由原暗黑模式装饰图案改造而来:两种模式都可用) ======
-// 共用一个渐变底 + 图案画笔:亮色用白色图案叠在主题色渐变上,暗色用主题色图案叠在黑底上。
-
-class _PatternSkin extends StatelessWidget {
-  const _PatternSkin(this.primary, this.isDark, this.painterFor);
-  final Color primary;
-  final bool isDark;
-  final CustomPainter Function(Color patternColor) painterFor;
-
-  @override
-  Widget build(BuildContext context) {
-    // 透明底:让 header 基础色透出(亮=主题色 / 暗=纯黑)。暗色模式下图案用主题色,
-    // 与原「暗黑模式头部图案」**完全一致**;亮色模式用白色让图案在主题色底上可见。
-    return CustomPaint(
-      painter: painterFor(isDark ? primary : Colors.white),
-      child: const SizedBox.expand(),
-    );
-  }
-}
-
-/// 蜂巢六边形
-class _HoneycombPainter extends CustomPainter {
-  _HoneycombPainter(this.color);
-  final Color color;
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    final paint = Paint()
-      ..color = color.withValues(alpha: 0.15)
-      ..style = PaintingStyle.stroke
-      ..strokeWidth = 1.5;
-    const hexSize = 30.0;
-    final hexHeight = hexSize * math.sqrt(3);
-    final hexWidth = hexSize * 2;
-    final rows = (size.height / hexHeight * 1.5).ceil() + 2;
-    final cols = (size.width / (hexWidth * 0.75)).ceil() + 2;
-    for (int row = -1; row < rows; row++) {
-      for (int col = -1; col < cols; col++) {
-        final x = col * hexWidth * 0.75;
-        final y = row * hexHeight + (col.isOdd ? hexHeight / 2 : 0);
-        final random = math.Random((row * 1000 + col).hashCode);
-        if (random.nextDouble() > 0.3) {
-          final path = Path();
-          for (int i = 0; i < 6; i++) {
-            final a = (math.pi / 3) * i;
-            final px = x + hexSize * math.cos(a);
-            final py = y + hexSize * math.sin(a);
-            if (i == 0) {
-              path.moveTo(px, py);
-            } else {
-              path.lineTo(px, py);
-            }
-          }
-          path.close();
-          canvas.drawPath(path, paint);
-        }
-      }
-    }
-  }
-
-  @override
-  bool shouldRepaint(covariant _HoneycombPainter old) => old.color != color;
-}
-
-/// 星河(粒子 + 五角星)
-class _StarryPainter extends CustomPainter {
-  _StarryPainter(this.color);
-  final Color color;
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    final random = math.Random(42); // 固定种子,保持一致性
-    for (int i = 0; i < 35; i++) {
-      final x = random.nextDouble() * size.width;
-      final y = random.nextDouble() * size.height;
-      final particleSize = 2.0 + random.nextDouble() * 4; // 2-6 px
-      final opacity = 0.1 + random.nextDouble() * 0.15; // 10%-25%
-      final paint = Paint()
-        ..color = color.withValues(alpha: opacity)
-        ..style = PaintingStyle.fill;
-      canvas.drawCircle(Offset(x, y), particleSize, paint);
-      // 20% 的粒子带光晕(与原暗黑装饰一致)
-      if (i % 5 == 0) {
-        final glowPaint = Paint()
-          ..color = color.withValues(alpha: opacity * 0.3)
-          ..style = PaintingStyle.fill
-          ..maskFilter = const MaskFilter.blur(BlurStyle.normal, 3);
-        canvas.drawCircle(Offset(x, y), particleSize * 2, glowPaint);
-      }
-    }
-    for (int i = 0; i < 10; i++) {
-      final x = random.nextDouble() * size.width;
-      final y = random.nextDouble() * size.height;
-      final starSize = 8.0 + random.nextDouble() * 8; // 8-16px
-      final opacity = 0.15 + random.nextDouble() * 0.1;
-      final paint = Paint()
-        ..color = color.withValues(alpha: opacity)
-        ..style = PaintingStyle.fill;
-      canvas.drawPath(_star(x, y, starSize, starSize * 0.4), paint);
-    }
-  }
-
-  Path _star(double cx, double cy, double outer, double inner) {
-    final path = Path();
-    const points = 5;
-    const angle = math.pi / points;
-    for (int i = 0; i < points * 2; i++) {
-      final radius = i.isEven ? outer : inner;
-      final a = angle * i - math.pi / 2;
-      final x = cx + radius * math.cos(a);
-      final y = cy + radius * math.sin(a);
-      if (i == 0) {
-        path.moveTo(x, y);
-      } else {
-        path.lineTo(x, y);
-      }
-    }
-    path.close();
-    return path;
-  }
-
-  @override
-  bool shouldRepaint(covariant _StarryPainter old) => old.color != color;
-}
-
-/// 斜纹(三组不同粗细的对角线)
-class _StripesPainter extends CustomPainter {
-  _StripesPainter(this.color);
-  final Color color;
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    const spacing = 30.0;
-    final paints = [
-      Paint()
-        ..color = color.withValues(alpha: 0.12)
-        ..style = PaintingStyle.stroke
-        ..strokeWidth = 2,
-      Paint()
-        ..color = color.withValues(alpha: 0.08)
-        ..style = PaintingStyle.stroke
-        ..strokeWidth = 1.5,
-      Paint()
-        ..color = color.withValues(alpha: 0.05)
-        ..style = PaintingStyle.stroke
-        ..strokeWidth = 1,
-    ];
-    for (int g = 0; g < 3; g++) {
-      for (double i = -size.height + spacing * g;
-          i < size.width + size.height;
-          i += spacing * 3) {
-        canvas.drawLine(
-            Offset(i, 0), Offset(i + size.height, size.height), paints[g]);
-      }
-    }
-  }
-
-  @override
-  bool shouldRepaint(covariant _StripesPainter old) => old.color != color;
-}
-
-// ====== 场景皮肤(代码绘制,跟随主题色)======
-
-class _SunsetSkin extends StatelessWidget {
-  const _SunsetSkin(this.primary, this.isDark);
-  final Color primary;
-  final bool isDark;
-
-  @override
-  Widget build(BuildContext context) {
-    final colors = isDark
-        ? [Colors.black, Colors.black]
-        : [_lighten(_hueShift(primary, 18), 0.22), _lighten(primary, 0.04)];
-    return DecoratedBox(
-      decoration: BoxDecoration(
-        gradient: LinearGradient(
-          begin: Alignment.topCenter,
-          end: Alignment.bottomCenter,
-          colors: colors,
-        ),
-      ),
-      child: CustomPaint(
-        painter: _SunsetPainter(primary, isDark),
-        child: const SizedBox.expand(),
-      ),
-    );
-  }
-}
-
-class _SunsetPainter extends CustomPainter {
-  _SunsetPainter(this.primary, this.isDark);
-  final Color primary;
-  final bool isDark;
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    final w = size.width, h = size.height;
-    // 落日
-    canvas.drawCircle(
-      Offset(w * 0.5, h * 0.66),
-      h * 0.3,
-      Paint()
-        ..color = (isDark ? primary : Colors.white)
-            .withValues(alpha: isDark ? 0.22 : 0.6),
-    );
-    // 远山
-    final back = Paint()
-      ..color = (isDark ? primary : _lighten(primary, -0.02))
-          .withValues(alpha: isDark ? 0.15 : 0.45);
-    final p1 = Path()
-      ..moveTo(0, h)
-      ..lineTo(0, h * 0.82);
-    p1.quadraticBezierTo(w * 0.28, h * 0.66, w * 0.55, h * 0.8);
-    p1.quadraticBezierTo(w * 0.8, h * 0.92, w, h * 0.74);
-    p1
-      ..lineTo(w, h)
-      ..close();
-    canvas.drawPath(p1, back);
-    // 近山
-    final front = Paint()
-      ..color = (isDark ? primary : _lighten(primary, -0.1))
-          .withValues(alpha: isDark ? 0.26 : 0.66);
-    final p2 = Path()
-      ..moveTo(0, h)
-      ..lineTo(0, h * 0.9);
-    p2.quadraticBezierTo(w * 0.35, h * 0.78, w * 0.62, h * 0.92);
-    p2.quadraticBezierTo(w * 0.85, h, w, h * 0.88);
-    p2
-      ..lineTo(w, h)
-      ..close();
-    canvas.drawPath(p2, front);
-  }
-
-  @override
-  bool shouldRepaint(covariant _SunsetPainter old) =>
-      old.primary != primary || old.isDark != isDark;
-}
-
-class _CloudsSkin extends StatelessWidget {
-  const _CloudsSkin(this.primary, this.isDark);
-  final Color primary;
-  final bool isDark;
-
-  @override
-  Widget build(BuildContext context) {
-    final colors = isDark
-        ? [Colors.black, Colors.black]
-        : [_lighten(primary, 0.22), _lighten(primary, 0.08)];
-    return DecoratedBox(
-      decoration: BoxDecoration(
-        gradient: LinearGradient(
-          begin: Alignment.topCenter,
-          end: Alignment.bottomCenter,
-          colors: colors,
-        ),
-      ),
-      child: CustomPaint(
-        painter: _CloudsPainter(primary, isDark),
-        child: const SizedBox.expand(),
-      ),
-    );
-  }
-}
-
-class _CloudsPainter extends CustomPainter {
-  _CloudsPainter(this.primary, this.isDark);
-  final Color primary;
-  final bool isDark;
-
-  void _cloud(Canvas canvas, Offset c, double s, Paint paint) {
-    canvas.drawCircle(Offset(c.dx, c.dy), s, paint);
-    canvas.drawCircle(Offset(c.dx + s * 0.9, c.dy + s * 0.15), s * 0.8, paint);
-    canvas.drawCircle(Offset(c.dx - s * 0.9, c.dy + s * 0.2), s * 0.7, paint);
-    canvas.drawCircle(Offset(c.dx + s * 0.1, c.dy + s * 0.5), s * 0.85, paint);
-    canvas.drawRRect(
-      RRect.fromRectAndRadius(
-        Rect.fromLTRB(
-            c.dx - s * 1.6, c.dy + s * 0.2, c.dx + s * 1.7, c.dy + s * 0.95),
-        Radius.circular(s),
-      ),
-      paint,
-    );
-  }
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    final w = size.width, h = size.height;
-    final paint = Paint()
-      ..color = (isDark ? primary : Colors.white)
-          .withValues(alpha: isDark ? 0.13 : 0.6);
-    _cloud(canvas, Offset(w * 0.22, h * 0.34), h * 0.12, paint);
-    _cloud(canvas, Offset(w * 0.72, h * 0.22), h * 0.16, paint);
-    _cloud(canvas, Offset(w * 0.85, h * 0.66), h * 0.1, paint);
-  }
-
-  @override
-  bool shouldRepaint(covariant _CloudsPainter old) =>
-      old.primary != primary || old.isDark != isDark;
-}
-
-// ====== 图片皮肤(SVG 素材,全幅铺满)======
-// 整幅 BoxFit.cover 铺满 header,与代码皮肤尺寸一致;底色仅作 SVG 透明区兜底
-// (亮=主题色浅染 / 暗=纯黑)。
-// themed=true:用 colorFilter(srcIn)把整幅 SVG 重新着色成主题色,明暗层次靠 SVG
-//   自身的 fill-opacity 表现(暗黑底为纯黑,主题色自然偏淡);themed=false(默认):
-//   按 SVG 自带的内联 fill 渲染(固定配色)。
-class _ImageSkin extends StatelessWidget {
-  const _ImageSkin(this.asset, this.primary, this.isDark, {this.themed = false});
-  final String asset;
-  final Color primary;
-  final bool isDark;
-  final bool themed;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      color: isDark ? Colors.black : _lighten(primary, 0.16),
-      child: SvgPicture.asset(
-        asset,
-        fit: BoxFit.cover,
-        alignment: Alignment.center,
-        colorFilter: themed ? ColorFilter.mode(primary, BlendMode.srcIn) : null,
-      ),
-    );
-  }
 }
